✅ test: admin 로그인, 회원가입, sessionId API 테스트 코드 작성
